+++ CHANGELOG.md
@@ -3,8 +3,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
+ 
+## Unreleased
+### Added
+- Board methods .getRows() and getColumns(). 
 
-## [Unreleased]
+## 1.0.0-alpha.1 - 2017-03-25
 ### Added
 - AttackEnemy action. 
 - BoardBuilder. 
@@ -15,7 +19,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Rename Player class to PlayerInterface.
 - Rename ReplenishArmy action to ReplenishGarrison. 
 - Rename BaseCastle action to BuildCastle.
-- Board constructor expects rows and columns. 
+- Board constructor expects rows and columns.
 
 ## 1.0.0-alpha - 2017-03-19
 ### Added